feat: update GitHub PR operation limits and add human comment filtering utility
